git on FreeBSD-8.1R amd64
Johannes Hanika (hanatos@gmail.com) [10.11.03 09:16] wrote:
> you probably need to take the ifdef APPLE branch in line 1021..?

sorry? ...

> or does it go away if you cast the dt_conf_get_int to (size_t) before
> the MIN MAX stuff?

mmm ... not sure

after:

+++ src/common/image.c
@@ -1020,9 +1020,9 @@ int dt_image_alloc(dt_image_t *img, dt_image_buffer_t mip)
   // clamped between 50MB -- 4GB
 #if defined(__APPLE__) || defined(__MACH__)
   // apple's size_t sucks
-  size_t max_mem = (size_t)(MAX(52428800, dt_conf_get_int("cache_memory"))/(float)DT_IMAGE_FULL);
+  size_t max_mem = (size_t)(MAX(52428800, (size_t)dt_conf_get_int("cache_memory"))/(float)DT_IMAGE_FULL);
 #else
-  size_t max_mem = (size_t)(MIN(4294967295, MAX(52428800, dt_conf_get_int("cache_memory")))/(float)DT_IMAGE_FULL);
+  size_t max_mem = (size_t)(MIN(4294967295, MAX(52428800, (size_t)dt_conf_get_int("cache_memory")))/(float)DT_IMAGE_FULL);
 #endif
   dt_print(DT_DEBUG_CACHE, "[image_alloc] mip %d uses %.3f/%.3f MB, alloc %.3f MB\n", mip, 
            cache->total_size[mip]/(1024.0*1024.0), max_mem/(1024.0*1024.0), size/(1024.0*1024.0));
